style: завершить разбивку на файлы компонентов и стилей
