+++ src/components/header/nav-link/nav-link.module.css
@@ -1,0 +1,9 @@
+.link {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  cursor: pointer;
+  border: none;
+  background: transparent;
+  border-radius: 40px;
+}
