+++ src/components/main/burger-constructor/burger-list-item/burger-list-item.js
@@ -1,4 +1,3 @@
-import React from "react";
 import PropTypes from "prop-types";
 import {
   ConstructorElement,
