+++ src/components/main/burger-constructor/price-box/price-box.js
@@ -1,4 +1,3 @@
-import React from "react";
 import PropTypes from "prop-types";
 import {
   CurrencyIcon,
