+++ src/components/main/burger-ingredients/burger-ingredients.module.css
@@ -1,55 +1,5 @@
 .ingredients {
-	display: flex;
-	flex-direction: column;
-	width: 100%;
-}
-
-.tabs {
-	width: 100%;
-	display: flex;
-}
-
-.rowsContainer {
-  max-height: calc(100vh - 88px - 156px - 40px - 24px);
-  /* 88 — header, 156 — margins + title, 40px - first top margin
-	24 — bottom from frame */
-  overflow-y: scroll;
-}
-
-.rowsContainer::-webkit-scrollbar {
-  width: 8px;
-  background: #2f2f37;
-}
-
-.rowsContainer::-webkit-scrollbar-thumb {
-  background: #8585ad;
-}
-
-.row:first-of-type {
-  margin-top: 0;
-}
-
-.list {
-  list-style: none;
-  margin-bottom: 0;
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  column-gap: 24px;
-  row-gap: 32px;
-}
-
-.item {
   display: flex;
   flex-direction: column;
-  row-gap: 4px;
-  justify-content: center;
-  text-align: center;
-  position: relative;
-}
-
-.price {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
+  width: 100%;
 }
