+++ src/components/main/burger-ingredients/ingredients/ingredient/ingredient.module.css
@@ -1,0 +1,15 @@
+.item {
+  display: flex;
+  flex-direction: column;
+  row-gap: 4px;
+  justify-content: center;
+  text-align: center;
+  position: relative;
+}
+
+.price {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+}
