+++ src/components/main/burger-ingredients/ingredients/ingredients-row/ingredients-row.module.css
@@ -1,0 +1,12 @@
+.row:first-of-type {
+  margin-top: 0;
+}
+
+.list {
+  list-style: none;
+  margin-bottom: 0;
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  column-gap: 24px;
+  row-gap: 32px;
+}
