+++ src/components/main/burger-ingredients/tabs/tabs.module.css
@@ -1,0 +1,4 @@
+.tabs {
+  width: 100%;
+  display: flex;
+}
